Move all the Milestone and award selection into MilestoneAwardSelector.
