+++ src/GameSetup.ts
@@ -1,68 +1,17 @@
-import {ELYSIUM_AWARDS, HELLAS_AWARDS, ORIGINAL_AWARDS, VENUS_AWARDS} from './awards/Awards';
 import {Board} from './boards/Board';
 import {BoardName} from './boards/BoardName';
 import {ElysiumBoard} from './boards/ElysiumBoard';
 import {Game, GameId, GameOptions} from './Game';
 import {HellasBoard} from './boards/HellasBoard';
-import {ELYSIUM_MILESTONES, HELLAS_MILESTONES, ORIGINAL_MILESTONES, VENUS_MILESTONES} from './milestones/Milestones';
 import {OriginalBoard} from './boards/OriginalBoard';
-import {RandomMAOptionType} from './RandomMAOptionType';
-import {MilestoneAwardSelector} from './MilestoneAwardSelector';
-import {IDrawnMilestonesAndAwards} from './IDrawnMilestonesAndAwards';
 import {Player} from './Player';
 import {Resources} from './Resources';
 import {ColonyName} from './colonies/ColonyName';
 import {Color} from './Color';
-import {AresSetup} from './ares/AresSetup';
 import {TileType} from './TileType';
 import {Random} from './Random';
 
 export class GameSetup {
-  public static chooseMilestonesAndAwards = function(gameOptions: GameOptions): IDrawnMilestonesAndAwards {
-    let drawnMilestonesAndAwards: IDrawnMilestonesAndAwards = {
-      milestones: [],
-      awards: [],
-    };
-
-    const includeVenus = gameOptions.venusNextExtension && gameOptions.includeVenusMA;
-    const requiredQty = includeVenus ? 6 : 5;
-
-    switch (gameOptions.randomMA) {
-    case RandomMAOptionType.NONE:
-      switch (gameOptions.boardName) {
-      case BoardName.ORIGINAL:
-        drawnMilestonesAndAwards.milestones.push(...ORIGINAL_MILESTONES);
-        drawnMilestonesAndAwards.awards.push(...ORIGINAL_AWARDS);
-        break;
-      case BoardName.HELLAS:
-        drawnMilestonesAndAwards.milestones.push(...HELLAS_MILESTONES);
-        drawnMilestonesAndAwards.awards.push(...HELLAS_AWARDS);
-        break;
-      case BoardName.ELYSIUM:
-        drawnMilestonesAndAwards.milestones.push(...ELYSIUM_MILESTONES);
-        drawnMilestonesAndAwards.awards.push(...ELYSIUM_AWARDS);
-        break;
-      }
-      if (includeVenus) {
-        drawnMilestonesAndAwards.milestones.push(...VENUS_MILESTONES);
-        drawnMilestonesAndAwards.awards.push(...VENUS_AWARDS);
-      }
-
-      break;
-    case RandomMAOptionType.LIMITED:
-      drawnMilestonesAndAwards = MilestoneAwardSelector.getRandomMilestonesAndAwards(gameOptions, requiredQty);
-      break;
-    case RandomMAOptionType.UNLIMITED:
-      drawnMilestonesAndAwards = MilestoneAwardSelector.getRandomMilestonesAndAwards(gameOptions, requiredQty, 100, 100, 100, 100);
-      break;
-    }
-
-    if (gameOptions.aresExtension) {
-      AresSetup.setupMilestonesAwards(drawnMilestonesAndAwards);
-    };
-    return drawnMilestonesAndAwards;
-  };
-
   // Function to construct the board and milestones/awards list
   public static newBoard(boardName: BoardName, shuffle: boolean, rng: Random, includeVenus: boolean): Board {
     if (boardName === BoardName.ELYSIUM) {
